adapt x264 style lambda tables [CHANGES OUTPUTS]
initLambda() in TComRom.cpp is passed the scale factor that is used to adjust
lambda2 values. It uses 0.85 right now, which is similar to x264's value (0.9),
and seems to be optimal for HEVC based on our tests.

These lambda tables seem to help most at low QPs, allowing the encoder to get
nearly lossless quality at high bitrates.