slice: move m_bReferenced to picsym and rename for clarity
This flag was being used for two purposes and introducing race hazards:

1. determining if a picture in the DPB must be signaled in an RPS
2. determining if the slice type is referenced (non-b)

The m_bHasReferences variable on the picsym is used for the first purpose.
The macro IS_REFERENCED(slice) is used for the second purpose.

The m_isReferenced member variable on the frame encoder was removed and
it just checks the slice type directly.