entropy: disable signaling of CABAC init state
This flag, which was already disabled when frame parallelism is in use (which
is nearly always) was of limited utility. It did not improve compression
efficiency by any measurable amount, and it was expensive to compute.  But the
quality which made it expendable was that it was the only user of the bBinsCoded
flag in the ContextModel; forcing us to copy twice as much data every time we
copy a context.

With this feature removed, the context model can be reduced to a single uint8_t
state variable.