entropy: give each Search instance its own set of RD contexts
This gives each ThreadLocalData a complete set of working contexts so each
thread can measure RD cost (for the same row) independent of one other. There
were content problems with the 'temp' and 'rqtRoot' and 'rqtTest' contexts.

For this to work we have to sync the 'cur' context to the slave prior
to it performing any RD measurements.

This commit finally removes the CI_IDX enums and uses a simple struct to hold
the contexts per depth; and the member variables were renamed from
"m_rdEntropyCoders" to "m_rdContexts" since these coders are only ever used to
save and restore CABAC state (never to code with)

This change exposed a bug. The next patch adds some tools to catch this class of
bug and the patch after that fixes it.