cmake: fix pkgconfig install destination
There is a problem with the install path of the package config file. As
suspected by me and best explained by user "BtbN" on Freenode IRC at #cmake,
"all install commands take the destination path relative to the install
prefix".

Thus, when generating the 'x265.pc' file, the $CMAKE_INSTALL_PREFIX is redundant
in the destination path and causes unwanted behaviour.