slicetype: fine tuning for slicetypeDecide batch work
At the start of the encode we have a full lookahead queue that all needs
analysis and slicetype planning and none of the frame encoders can start until
this work is finished, so we dedicate the entire thread pool towards this
effort.

But after this burst of activity, the lookahead is competing with the frame
encoders for worker thread attention and so it generally doesn't want to do any
batch work that might be unnecesary. So based on the thread pool size we
throttle down the work distribution in slicetypeDecide.