predict: introduce PredictionUnit structure to maintain current PU geometry
The whole idea of prepMotionCompensation was to cache some data that could be
shared between motionCompensation() calls but since the motion vectors and refs
were included in that data, it had to be re-done each time any way.  In the
mean-time, much code was taking advantage of the PU data members but in a
fragile way, data hazards were too common.

The PredictionUnit struct is intended to be stack allocated and the PU geometry
is valid so long as the structure is within scope, making the data safe.