intra: pull the simple 1:2:1 pixel filtering into a performance primitive
Only C-refs at this point, but at least it is templated so the compiler can
optimize and unroll loops cleanly.

This commit also changes the lookahead to use the full available reference
samples. Since we're using padded downscaled source images, the top-left and
bottom right reference samples are always available. There is no good reason
not to use them.