Performance: Prevent small thread-pools if NUMA disabled and # CPUs > MAX_POOL_THREADS
When NUMA is disabled and if # CPUs is > MAX_POOL_THREADS (64 or 32 depending on 64-bit
or 32-bit builds), the last pool may have few threads. This patch allows the last pool to
exist only if it has > MAX_POOL_THREADS/2 threads; the 50% number is a heurstic.

Feature gains performance of 5% on Intel Xeon E5-2699v3 measured in slower preset with 4K video.