perf: Enabling lookahead-slices for all presets except veryslow & placebo
Seeing ~10% performance on the faster presets on skylake, and ~2X performance
on Xeon systems in ultrafast setting. Performance improvement is through a
considerable increase in utilization across the board. Disabling slicing for
videos of resoultion < 720p to limit impact on quality.

Commit will change outputs, but reduction in quality (measured by PSNR or SSIM)
is <0.01% across a wide variety of presets and runs.