Clean up dynamic refinement
This patch does the following:
1) Earlier, locks were used to avoid the possibility of race conditions while
copying data from CTU level to frame level. Now, the data is collected for each
row and when the entire frame completes analysis the row data is copied to the
frame. This method eliminates the possibility of a race condition without
having to employ locks.
2) Allocate memory for the CTU infromation from the data pool, this will avoid
fragmentation of data.